+++ HISTORY.rst
@@ -12,6 +12,7 @@ HISTORY
 * remove unused modules from metmetpy
 * adding tests to cdm_reader_mapper testing suite
 * cdm_reader_mapper.mdf_reader split data_models into code_tables and schema
+* move converters and decoders from common to mdf_reader/utils
 
 0.1.0 (2023-01-16)
 ------------------